+++ src/types-internal/editor-modules.d.ts
@@ -5,7 +5,7 @@ import InlineToolbar from '../components/modules/toolbar/inline';
 import Toolbox from '../components/modules/toolbar/toolbox';
 import BlockSettings from '../components/modules/toolbar/blockSettings';
 import Events from '../components/modules/events';
-import Shortcuts from '../components/modules/shortcuts';
+import Shortcuts from '../components/utils/shortcuts';
 import Paste from '../components/modules/paste';
 import Notifier from '../components/modules/notifier';
 import Tooltip from '../components/modules/tooltip';