+++ docs/CHANGELOG.md
@@ -17,6 +17,7 @@
 - `Refactoring` - The Listeners module now is a util.
 - `Fix` - Editor Config now immutable [#1552](https://github.com/codex-team/editor.js/issues/1552).
 - `Refactoring` - The Sanitizer module is util now.
+- `Refactoring` - Shortcuts module is util now
 
 ### 2.19.1